Verify JWT tokens
Adds middleware to verify JWT tokens.
