+++ package.json
@@ -13,22 +13,24 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "aws-sdk": "^2.259.1",
+    "aws-sdk": "^2.441.0",
     "bunyan": "^1.8.12",
     "bunyan-middleware": "^0.8.0",
-    "compression": "^1.7.2",
-    "cors": "^2.8.4",
-    "dotenv": "^6.0.0",
-    "express": "^4.16.3",
-    "express-graphql": "^0.6.12",
-    "graphql": "^0.13.2",
-    "helmet": "^3.12.1"
+    "compression": "^1.7.4",
+    "cors": "^2.8.5",
+    "dotenv": "^7.0.0",
+    "express": "^4.16.4",
+    "express-graphql": "^0.8.0",
+    "express-jwt": "^5.3.1",
+    "graphql": "^14.2.1",
+    "helmet": "^3.16.0",
+    "jsonwebtoken": "^8.5.1"
   },
   "devDependencies": {
-    "chai": "^4.1.2",
-    "eslint": "^4.19.1",
+    "chai": "^4.2.0",
+    "eslint": "^5.16.0",
     "fakefile": "0.0.9",
-    "mocha": "^5.2.0",
-    "supertest": "^3.1.0"
+    "mocha": "^6.1.4",
+    "supertest": "^4.0.2"
   }
 }
