+++ src/index.js
@@ -1,4 +1,5 @@
 "use strict";
+require("dotenv").config({silent: true});
 const app = require("./app");
 const port = process.env.PORT || 5000;
